Updated code implementation
