+++ app/src/main/java/org/oppia/app/drawer/NavigationDrawerFragmentPresenter.kt
@@ -7,12 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.widget.Toolbar
-import androidx.databinding.DataBindingUtil
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.Observer
-import androidx.lifecycle.Transformations
 import com.google.android.material.navigation.NavigationView
 import org.oppia.app.R
 import org.oppia.app.databinding.DrawerFragmentBinding
@@ -20,63 +16,33 @@ import org.oppia.app.databinding.NavHeaderNavigationDrawerBinding
 import org.oppia.app.fragment.FragmentScope
 import org.oppia.app.help.HelpActivity
 import org.oppia.app.home.HomeActivity
-import org.oppia.app.model.Profile
-import org.oppia.app.model.ProfileId
-import org.oppia.domain.profile.ProfileManagementController
-import org.oppia.util.data.AsyncResult
-import org.oppia.util.logging.Logger
+import org.oppia.app.viewmodel.ViewModelProvider
 import javax.inject.Inject
 
 /** The presenter for [NavigationDrawerFragment]. */
 @FragmentScope
 class NavigationDrawerFragmentPresenter @Inject constructor(
   private val fragment: Fragment,
-  private val profileManagementController: ProfileManagementController,
-  private val logger: Logger
+  private val headerViewModelProvider: ViewModelProvider<NavigationDrawerHeaderViewModel>
 ) : NavigationView.OnNavigationItemSelectedListener {
-  private lateinit var navView: NavigationView
   private lateinit var drawerToggle: ActionBarDrawerToggle
   private lateinit var drawerLayout: DrawerLayout
   private var previousMenuItemId: Int? = null
   private lateinit var binding: DrawerFragmentBinding
-  private lateinit var navHeaderMainBinding: NavHeaderNavigationDrawerBinding
-  private lateinit var profileId: ProfileId
 
   fun handleCreateView(inflater: LayoutInflater, container: ViewGroup?): View? {
     binding = DrawerFragmentBinding.inflate(inflater, container, /* attachToRoot= */ false)
-    navView = binding.fragmentDrawerNavView
     binding.fragmentDrawerNavView.setNavigationItemSelectedListener(this)
 
     fragment.setHasOptionsMenu(true)
-    profileId = profileManagementController.getCurrentProfileId()
-    navHeaderMainBinding =
-      DataBindingUtil.inflate(inflater, R.layout.nav_header_navigation_drawer, navView, true)
-    navView.setNavigationItemSelectedListener(this)
+    binding.fragmentDrawerNavView.setNavigationItemSelectedListener(this)
 
-    subscribeToProfileLiveData()
+    val headerBinding = NavHeaderNavigationDrawerBinding.inflate(inflater, container, /* attachToRoot= */ false)
+    headerBinding.viewModel = getNavigationDrawerHeaderViewModel()
 
+    binding.fragmentDrawerNavView.addHeaderView(headerBinding.root)
+    binding.executePendingBindings()
     return binding.root
-  }
-
-  private val profileLiveData: LiveData<Profile> by lazy {
-    getProfileData()
-  }
-
-  private fun getProfileData(): LiveData<Profile> {
-    return Transformations.map(profileManagementController.getProfile(profileId), ::processGetProfileResult)
-  }
-
-  private fun subscribeToProfileLiveData() {
-    profileLiveData.observe(fragment, Observer<Profile> { result ->
-      navHeaderMainBinding.navNameTextview.text = result.name
-    })
-  }
-
-  private fun processGetProfileResult(profileResult: AsyncResult<Profile>): Profile {
-    if (profileResult.isFailure()) {
-      logger.e("NavigationDrawerFragmentPresenter", "Failed to retrieve profile", profileResult.getErrorOrNull()!!)
-    }
-    return profileResult.getOrDefault(Profile.getDefaultInstance())
   }
 
   private fun openActivityByMenuItemId(menuItemId: Int) {
@@ -103,10 +69,10 @@ class NavigationDrawerFragmentPresenter @Inject constructor(
   fun setUpDrawer(drawerLayout: DrawerLayout, toolbar: Toolbar, menuItemId: Int) {
     when (NavigationDrawerItem.valueFromNavId(menuItemId)) {
       NavigationDrawerItem.HOME -> {
-        navView.menu.getItem(NavigationDrawerItem.HOME.ordinal).isChecked = true
+        binding.fragmentDrawerNavView.menu.getItem(NavigationDrawerItem.HOME.ordinal).isChecked = true
       }
       NavigationDrawerItem.HELP -> {
-        navView.menu.getItem(NavigationDrawerItem.HELP.ordinal).isChecked = true
+        binding.fragmentDrawerNavView.menu.getItem(NavigationDrawerItem.HELP.ordinal).isChecked = true
       }
     }
     this.drawerLayout = drawerLayout
@@ -136,5 +102,9 @@ class NavigationDrawerFragmentPresenter @Inject constructor(
   override fun onNavigationItemSelected(menuItem: MenuItem): Boolean {
     openActivityByMenuItemId(menuItem.itemId)
     return true
+  }
+
+  private fun getNavigationDrawerHeaderViewModel(): NavigationDrawerHeaderViewModel {
+    return headerViewModelProvider.getForFragment(fragment, NavigationDrawerHeaderViewModel::class.java)
   }
 }
