+++ app/src/main/java/org/oppia/app/topic/conceptcard/ConceptCardViewModel.kt
@@ -1,6 +1,5 @@
 package org.oppia.app.topic.conceptcard
 
-import android.widget.TextView
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
