+++ app/src/main/res/layout/drawer_fragment.xml
@@ -8,7 +8,6 @@
     android:layout_height="match_parent"
     android:layout_gravity="start"
     android:fitsSystemWindows="true"
-    app:headerLayout="@layout/nav_header_navigation_drawer"
     app:itemBackground="@android:color/transparent"
     app:itemIconTint="@color/drawer_item"
     app:itemTextColor="@color/drawer_item"
